move irrep functions out of common.utils
