+++ src/fairchem/core/models/uma/escn_md.py
@@ -17,7 +17,7 @@ from torch.profiler import record_function
 from fairchem.core.common import gp_utils
 from fairchem.core.common.distutils import get_device_for_local_rank
 from fairchem.core.common.registry import registry
-from fairchem.core.common.utils import cg_change_mat, conditional_grad, irreps_sum
+from fairchem.core.common.utils import conditional_grad
 from fairchem.core.graph.compute import generate_graph
 from fairchem.core.models.base import HeadInterface
 from fairchem.core.models.uma.common.rotation import (
@@ -41,6 +41,7 @@ from fairchem.core.models.uma.nn.layer_norm import (
 from fairchem.core.models.uma.nn.mole_utils import MOLEInterface
 from fairchem.core.models.uma.nn.radial import GaussianSmearing
 from fairchem.core.models.uma.nn.so3_layers import SO3_Linear
+from fairchem.core.models.utils.irreps import cg_change_mat, irreps_sum
 
 from .escn_md_block import eSCNMD_Block
