+++ src/fairchem/core/modules/transforms.py
@@ -11,7 +11,7 @@ from typing import TYPE_CHECKING
 
 import torch
 
-from fairchem.core.common.utils import cg_change_mat, irreps_sum
+from fairchem.core.models.utils.irreps import cg_change_mat, irreps_sum
 
 if TYPE_CHECKING:
     from torch_geometric.data import Data
